+++ source/dataparsing/correction.py
@@ -4,37 +4,28 @@ __author__ = "Guy Keogh"
 __license__ = "BSD 2-Clause"
 """
 
-from source import text_tagging
+from source.dataparsing import text_tagging
 
 def add_input(input_text, data, settings):
-    each_URL_location = []
     processed_citations = data['processed_citations']
-    #print("Processed citations: " + str(processed_citations))
     URL_indexes = dict()
 
     for external_URL in processed_citations:
         if not external_URL or external_URL == '':
             continue
 
-        print("Checking " + external_URL)
         position = input_text.find(external_URL)
         if position != -1: # If some URL on the page is within the user input
             URL_indexes.update({position+len(external_URL): external_URL}) # Save the start position of the URL and the URL itself
-            print("Found URL. New index: " + str(URL_indexes))
-        else:
-            print("Nope.")
     
     if not URL_indexes: #No URLs supplied
         data['reprocess?'] = False
-        print("No urls given")
     else:
         # Order the indexes:
         URL_indexes_ordered = []
         for index in URL_indexes:
             URL_indexes_ordered.append(index) # Get the index from the dictionary with structure {index: URL}
         URL_indexes_ordered = sorted(URL_indexes_ordered)
-
-        print("ordered indexes: " + str(URL_indexes_ordered))
 
         #Add the inputted text to its value in the dictionary of citations and their data:
         indexes_index = 0
@@ -52,8 +43,6 @@ def add_input(input_text, data, settings):
                 position_end = URL_indexes_ordered[indexes_index+1] # Start of next URL in list
             else:
                 position_end = len(input_text) #End of inputted text - no more URLs
-            
-            print("Start is " + str(position_start) + " and end is " + str(position_end))
             # Using the start and end indexes, find the relevant text:
             text = input_text[position_start:position_end]
 
@@ -62,7 +51,6 @@ def add_input(input_text, data, settings):
 
             # Finally, do all the processing, adding the user-inputted text to the processed citations:
             citation_words = text_tagging.tag_citation_text(citation_words, text, settings)
-            print("Added " + str(citation_words) + " to " + external_URL)
             processed_citations[external_URL] = citation_words
             indexes_index+=1
